refactor: fix last deprecations
