+++ phpstan-baseline.neon
@@ -41,6 +41,21 @@ parameters:
 			path: src/FactoryCollection.php
 
 		-
+			message: "#^Class Zenstruck\\\\Foundry\\\\Instantiator extends @final class Zenstruck\\\\Foundry\\\\Object\\\\Instantiator\\.$#"
+			count: 1
+			path: src/Instantiator.php
+
+		-
+			message: "#^If condition is always false\\.$#"
+			count: 1
+			path: src/Instantiator.php
+
+		-
+			message: "#^Class Zenstruck\\\\Foundry\\\\Object\\\\ObjectFactory not found\\.$#"
+			count: 2
+			path: src/ModelFactory.php
+
+		-
 			message: "#^Method Zenstruck\\\\Foundry\\\\Persistence\\\\PersistentObjectFactory\\:\\:__callStatic\\(\\) should return array\\<int, TModel of object\\> but returns array\\<int, \\(TModel of object\\)\\|Zenstruck\\\\Foundry\\\\Persistence\\\\Proxy\\<TModel of object\\>\\>\\.$#"
 			count: 1
 			path: src/Persistence/PersistentObjectFactory.php
@@ -81,9 +96,19 @@ parameters:
 			path: src/Proxy.php
 
 		-
+			message: "#^Class Zenstruck\\\\Foundry\\\\RepositoryAssertions extends @final class Zenstruck\\\\Foundry\\\\Persistence\\\\RepositoryAssertions\\.$#"
+			count: 1
+			path: src/RepositoryAssertions.php
+
+		-
 			message: "#^If condition is always false\\.$#"
 			count: 1
 			path: src/RepositoryAssertions.php
+
+		-
+			message: "#^Class Zenstruck\\\\Foundry\\\\RepositoryProxy extends @final class Zenstruck\\\\Foundry\\\\Persistence\\\\RepositoryDecorator\\.$#"
+			count: 1
+			path: src/RepositoryProxy.php
 
 		-
 			message: "#^If condition is always false\\.$#"
