+++ src/Persistence/PersistentProxyObjectFactory.php
@@ -44,7 +44,9 @@ abstract class PersistentProxyObjectFactory extends PersistentObjectFactory
         if ((new \ReflectionClass(static::class()))->isFinal()) {
             trigger_deprecation(
                 'zenstruck\foundry', '1.37.0',
-                \sprintf('Using a proxy factory with a final class is deprecated and will throw an error in Foundry 2.0. Use "%s" instead, or unfinalize "%s" class.', self::class, static::class())
+                'Using a proxy factory with a final class is deprecated and will throw an error in Foundry 2.0. Use "Zenstruck\Foundry\Object\ObjectFactory" instead (don\'t forget to remove all ->object() calls!).',
+                self::class,
+                static::class()
             );
         }
 
@@ -63,7 +65,7 @@ abstract class PersistentProxyObjectFactory extends PersistentObjectFactory
         bool $noProxy = false
     ): object {
         if (2 === \count(\func_get_args()) && !\str_starts_with(\debug_backtrace(options: \DEBUG_BACKTRACE_IGNORE_ARGS, limit: 1)[0]['class'] ?? '', 'Zenstruck\Foundry')) {
-            trigger_deprecation('zenstruck\foundry', '1.37.0', \sprintf('Parameter "$noProxy" of method "%s()" is deprecated and will be removed in Foundry 2.0.', __METHOD__));
+            trigger_deprecation('zenstruck\foundry', '1.37.0', 'Parameter "$noProxy" of method "%s()" is deprecated and will be removed in Foundry 2.0.', __METHOD__);
         }
 
         return Factory::create($attributes, noProxy: false);
