+++ src/RepositoryAssertions.php
@@ -13,13 +13,23 @@ namespace Zenstruck\Foundry;
 
 use Zenstruck\Foundry\Persistence\RepositoryAssertions as NewRepositoryAssertions;
 
-trigger_deprecation('zenstruck\foundry', '1.37.0', \sprintf('Class "%s" is deprecated and will be removed in version 2.0. Use "%s" instead.', RepositoryAssertions::class, NewRepositoryAssertions::class));
+if (!class_exists(NewRepositoryAssertions::class, false)) {
+    trigger_deprecation(
+        'zenstruck\foundry',
+        '1.37.0',
+        'Class "%s" is deprecated and will be removed in version 2.0. Use "%s" instead.',
+        RepositoryAssertions::class,
+        NewRepositoryAssertions::class
+    );
+}
+
+\class_alias(NewRepositoryAssertions::class, RepositoryAssertions::class);
 
 if (false) {
     /**
      * @deprecated
      */
-    final class RepositoryAssertions
+    final class RepositoryAssertions extends NewRepositoryAssertions
     {
     }
 }
